Fixed extra space to right of search button
The search bar has had its width reduced slightly so that there is no gap between the search button and the box shadow. The height of each child has been set with `height: 100%` so that they will inherit the `38px` height. Additionally, the default padding and margin has been removed from the button, the arrow background no longer repeats, and is now perfectly centered using background position `11px` for the y axis and `50%` for the x axis.